Fixes to Ladspa plugin effects that make avoid mono/stereo and other crashes. Patch from Nicholas Lamb.

+++ sf-cvs/trunk/audacity-src/src/effects/ladspa/LadspaEffect.h
@@ -47,9 +47,9 @@ class LadspaEffect:public Effect {
    float *buffer;
    float **fInBuffer;
    float **fOutBuffer;
-   int inputs;
-   int outputs;
-   int numInputControls;
+   unsigned long inputs;
+   unsigned long outputs;
+   unsigned long numInputControls;
    unsigned long *inputPorts;
    unsigned long *outputPorts;
    float *inputControls;
